+++ common/src/main/java/net/wizards/entity/ArcaneEmitterEntity.java
@@ -1,14 +1,10 @@
 package net.wizards.entity;
 
 import net.minecraft.entity.EntityType;
-import net.minecraft.util.Identifier;
 import net.minecraft.world.World;
 import net.spell_engine.entity.SummonedEntity;
-import net.wizards.WizardsMod;
 
 public class ArcaneEmitterEntity extends SummonedEntity {
-    public static final Identifier ID = Identifier.of(WizardsMod.ID, "arcane_emitter");
-    public static EntityType<ArcaneEmitterEntity> TYPE;
 
     public ArcaneEmitterEntity(EntityType<? extends ArcaneEmitterEntity> entityType, World world) {
         super(entityType, world);